LayoutTests/crypto/subtle/aes-cbc-wrap-rsa-non-extractable-expected.txt

Test wrapping an RSA key with AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key to wrap...
PASS key.algorithm.name is 'rsassa-pkcs1-v1_5'
Importing a key encryption key...
PASS wrappingKey.algorithm.name is 'aes-cbc'
wrapping the key...
PASS Rejected
PASS successfullyParsed is true

TEST COMPLETE
